removed duplicated code, reorganized nav images, renamed dashboard
