+++ src/components/dashboardComponents/CreateTemplateForm.js
@@ -2,25 +2,10 @@ import React from 'react';
 import { useForm } from 'react-hook-form';
 import { Button, Input } from '@chakra-ui/core';
 import axios from 'axios';
-
-
-const addTemplate = async (data, { googleId }) => {
-  const template = { ...data, googleId };
-  try {
-    const response = await axios.post(
-      `${process.env.REACT_APP_ENDPOINT_URL}/api/template`,
-      template
-    );
-    // console.log('response.data', response.data)
-    return response.data;
-  } catch (error) {
-    console.log(error);
-  }
-};
+import {addTemplate} from '../../utils/helperFunctions'
 
 const CreateTemplateForm = (props) => {
   const { setFormOpen, setTemplateList, currentUser, formOpen } = props;
-  // const { googleApi, api } = useAuth();
   const { register, handleSubmit } = useForm();
 
   // Submit for template form
