+++ src/components/dashboardComponents/calendarComponents/EventsIndicator.js
@@ -1,34 +1,19 @@
 import React from 'react';
 import dayjs from 'dayjs';
 import styled from 'styled-components';
-
 import { Icon } from '@chakra-ui/core';
-
-const Event = styled.div`
-  font-size: 1.25rem;
-`
 
 // Displays circle icon if day has an event
 const EventIndicator = ({ event, day, currentMonth, currentYear }) => {
-  // const hasEvents =
-  //   day === dayjs(event.start.dateTime).date() &&
-  //   currentMonth === dayjs(event.start.dateTime).month() &&
-  //   currentYear === dayjs(event.start.dateTime).year();
 
   let hasEvents;
-  if(event.start.dateTime){
+  if(event.start.dateTime || event.start.date){
     hasEvents =
     day === dayjs(event.start.dateTime).date() &&
     currentMonth === dayjs(event.start.dateTime).month() &&
     currentYear === dayjs(event.start.dateTime).year();
-  } else if (event.start.date){
-    hasEvents =
-    day === dayjs(event.start.date).date() &&
-    currentMonth === dayjs(event.start.date).month() &&
-    currentYear === dayjs(event.start.date).year();
-  }
+  } 
   
-
   return hasEvents ? (
     <>
       <Icon name="circle" fill="brand.blue_accent" />
@@ -38,3 +23,7 @@ const EventIndicator = ({ event, day, currentMonth, currentYear }) => {
 };
 
 export default EventIndicator;
+
+const Event = styled.div`
+  font-size: 1.25rem;
+`
